feat: 🎸 Disabled application form
Also updated a bunch of dependencies
