+++ .babelrc
@@ -1,10 +1,4 @@
 {
-  "presets": [
-    "@babel/preset-env",
-    "@babel/react",
-    "next/babel"
-  ],
-  "plugins": [
-    "@babel/plugin-transform-runtime"
-  ]
+  "presets": ["@babel/preset-env", "@babel/react", "next/babel"],
+  "plugins": ["@babel/plugin-transform-runtime"]
 }
