--- .eslintrc
@@ -1,13 +1,0 @@
-{
-  "extends": ["plugin:cypress/recommended", "plugin:react/jsx-runtime"],
-  "parser": "@babel/eslint-parser",
-  "ignorePatterns": ["/*.js"],
-  "parserOptions": {
-    "sourceType": "module",
-    "ecmaVersion": 2020
-  },
-  "rules": {
-    "react/react-in-jsx-scope": 0,
-    "react/prop-types": 0
-  }
-}
