+++ content/intresse_open.md
@@ -1,0 +1,16 @@
+---
+name: Intresseanmälan
+type: Prospect
+title: Intresseanmälan
+
+headline: Intresseanmälan
+---
+
+Vår koloniförening har bara totalt 27 lotter/stugor och vi har för närvarande flera hundratals anmälda till vår intresselista. Omsättningen på lotter är i dagsläget mycket låg. Vill du ändå anmäla ditt intresse för en kolonilott i Karlbergs-Bro Koloniförening så kan du anmäla dig här. Observera att det endast är personer som är mantalsskrivna inom Storstockholm som kan antas som medlemmar i föreningen.
+
+Vår intresselista är inte en kö i vanlig mening. Då en lott blir ledig kontaktar föreningens styrelse ett antal personer i kölistan för intervju. Urval sker utifrån lämplighet och egenskaper som föreningen söker hos en ny lottägare.
+Eftersom omsättningen av lotter är låg och intresselistan inte fungerar som en vanlig kölista kan vi inte svara på frågor som handlar om hur lång kötid det är för en kolonilott i föreningen.
+
+Din anmälan registreras först när du betalt 200:- i administrationsavgift till plusgiro 3 27 69 – 2 (Karlbergs-Bro Koloniförening, Stockholm). Detta är en engångsavgift och du behöver inte betala någon fortlöpande köavgift.
+
+Var noga med att fylla i alla uppgifter, så att vi kan göra en korrekt bedömning. Det ligger på ditt ansvar att uppdatera adress, telefon samt e-postadress så att dina korrekta uppgifter finns i intresselistan.
