+++ package.json
@@ -10,58 +10,61 @@
     "format": "prettier --write --ignore-path .gitignore ./",
     "test": "jest",
     "pree2e": "CYPRESS_DIST=.e2e yarn build",
-    "e2e": "start-server-and-test 'CYPRESS_DIST=.e2e NODE_ENV=test yarn start -p 4000' :4000 'CYPRESS_BASE_URL=http://localhost:4000 yarn cy:run --config video=false'",
+    "e2e": "start-server-and-test 'CYPRESS_DIST=.e2e NODE_ENV=test yarn start -p 4444' :4444 'CYPRESS_BASE_URL=http://localhost:4444 yarn cy:run --config video=false'",
     "cy:open": "cypress open",
     "cy:run": "cypress run",
     "prepare": "husky install"
   },
   "dependencies": {
-    "bootstrap": "4.5.3",
+    "bootstrap": "4.6.2",
     "fast-memoize": "2.5.2",
     "gray-matter": "4.0.3",
-    "next": "13.1.1",
-    "nodemailer": "6.6.3",
+    "next": "13.5.6",
+    "nodemailer": "6.9.14",
     "normalize-unicode-text": "^0.0.1",
-    "react": "18.2.0",
-    "react-bootstrap": "1.6.1",
-    "react-dom": "18.2.0",
-    "react-markdown": "5.0.3",
-    "remark-gfm": "1.0.0"
+    "react": "18.3.1",
+    "react-bootstrap": "2.10.4",
+    "react-dom": "18.3.1",
+    "react-markdown": "9.0.1",
+    "remark-gfm": "4.0.0"
   },
   "devDependencies": {
-    "@babel/core": "^7.20.12",
-    "@babel/eslint-parser": "^7.19.1",
-    "@babel/plugin-transform-runtime": "^7.19.6",
-    "@babel/preset-env": "^7.20.2",
-    "@babel/preset-react": "^7.18.6",
-    "@testing-library/react": "^13.4.0",
-    "@types/jest": "^29.2.5",
-    "@types/node": "^18.11.18",
-    "@types/react": "^18.0.26",
-    "@typescript-eslint/eslint-plugin": "^5.48.0",
-    "babel-jest": "^29.3.1",
-    "cypress": "^12.3.0",
-    "eslint": "^8.31.0",
+    "@babel/core": "^7.25.2",
+    "@babel/eslint-parser": "^7.25.1",
+    "@babel/plugin-transform-runtime": "^7.25.4",
+    "@babel/preset-env": "^7.25.4",
+    "@babel/preset-react": "^7.24.7",
+    "@eslint/eslintrc": "^3.1.0",
+    "@eslint/js": "^9.9.1",
+    "@testing-library/react": "^16.0.0",
+    "@types/jest": "^29.5.12",
+    "@types/node": "^22.5.0",
+    "@types/react": "^18.3.4",
+    "@typescript-eslint/eslint-plugin": "^8.2.0",
+    "babel-jest": "^29.7.0",
+    "cypress": "^13.13.3",
+    "eslint": "^9.9.1",
     "eslint-config-airbnb": "^19.0.4",
-    "eslint-config-prettier": "^8.6.0",
-    "eslint-plugin-cypress": "^2.12.1",
-    "eslint-plugin-import": "^2.26.0",
-    "eslint-plugin-jsx-a11y": "^6.6.1",
-    "eslint-plugin-prettier": "^4.2.1",
-    "eslint-plugin-react": "^7.31.11",
-    "eslint-plugin-react-hooks": "^4.6.0",
+    "eslint-config-prettier": "^9.1.0",
+    "eslint-plugin-cypress": "^3.5.0",
+    "eslint-plugin-import": "^2.29.1",
+    "eslint-plugin-jsx-a11y": "^6.9.0",
+    "eslint-plugin-prettier": "^5.2.1",
+    "eslint-plugin-react": "^7.35.0",
+    "eslint-plugin-react-hooks": "^4.6.2",
     "git-cz": "^4.9.0",
-    "husky": "^8.0.3",
-    "jest": "^29.3.1",
+    "husky": "^9.1.5",
+    "jest": "^29.7.0",
     "mock-fs": "^5.2.0",
-    "prettier": "^2.8.2",
-    "sass": "^1.57.1",
-    "start-server-and-test": "^1.15.2",
-    "yarn": "^1.22.19"
+    "prettier": "^3.3.3",
+    "sass": "^1.77.8",
+    "start-server-and-test": "^2.0.5",
+    "yarn": "^1.22.22"
   },
   "config": {
     "commitizen": {
       "path": "./node_modules/git-cz"
     }
-  }
+  },
+  "packageManager": "yarn@4.4.1"
 }
