+++ pages/[[...page]].js
@@ -1,4 +1,3 @@
-/* eslint-disable react/no-danger */
 import Head from 'next/head'
 import {
   getPath,
